+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [1.0.0] / 2023-12-11
 - First Release
+- Use `Form` with `DoEvent` (`DoEvent` is faster.)
+- Disable `RevitRibbonControl` when `RevitProgressBar` running.
 
 [vNext]: ../../compare/1.0.0...HEAD
 [1.0.0]: ../../compare/1.0.0